requirements.txt
cython>=0.28.5
imageio>=2.6.1
matplotlib>=3.3.0
mxnet>=1.7.0.post1
numpy>=1.19.4
opencv-python>=4.3.0.36
pandas>=1.0.5
Pillow>=8.1.0
pydensecrf>=1.0rc3
pytest
scikit-image>=0.17.2
scipy>=1.5.1
tensorboard>=1.8.0
tensorboardX>=2.1
torch==1.9.0+cu111
torchvision==0.10.0+cu111
torchaudio==0.9.0
tqdm
